Define a more explicit `File_Contents` check
This means we can also have a `File_Type` check
